Refactors field py files, adds featured module and type b module
