+++ navigation/fields/fields.py
@@ -1,0 +1,23 @@
+from wagtail.core.fields import StreamField
+from wagtail.core.blocks import (
+    StreamBlock,
+)
+from navigation.fields import (
+    TypeA,
+    TypeB,
+)
+
+
+def navigation(blank=False):
+    required = not blank
+    return StreamField(
+        StreamBlock(
+            [
+                ('type_a', TypeA()),
+                ('type_b', TypeB()),
+            ],
+            max_num=1,
+            required=required,
+        ),
+        blank=blank
+    )
