+++ navigation/fields/page.py
@@ -1,0 +1,15 @@
+from wagtail.core.blocks import (
+    PageChooserBlock,
+    StructBlock,
+)
+from navigation.values import TransStructValue
+
+
+class TranslatedPage(StructBlock):
+
+    class Meta:
+        icon = 'link'
+        label = 'Page'
+        value_class = TransStructValue
+
+    page = PageChooserBlock()
