+++ navigation/templates/navigation/blocks/featured.html
@@ -1,0 +1,23 @@
+ {% load i18n static %}
+
+<div class="l-offset__col">
+    <article class="case-study">
+        <div class="case-study__media background-cover" style="background-image: url({% static 'img/content-image.jpg' %}"></div>
+        <div class="case-study__panel">
+            <h2 class="case-study__heading">
+                <a href="{{ value.page.url }}"><span>{{ value.page.specific.heading }}</span></a>
+            </h2>
+            <p class="case-study__excerpt">{{ value.description }}</p>
+            <a href="{ value.page.url }}" class="button button--small">{% trans 'Read more' %}</a>
+            {% if value.secondary_page %}
+                <a href="{{ value.secondary_page.url }}" class="case-study__text-link">
+                    {% if value.link_label %}
+                        {{ value.link_label }}
+                    {% else %}
+                        {{ value.secondary_page.specific.heading }}
+                    {% endif %}
+                </a>
+            {% endif %}
+        </div>
+    </article>
+</div>
