+++ navigation/templates/navigation/blocks/highlight.html
@@ -1,4 +1,4 @@
- {% load wagtail_modeltranslation i18n %}
+ {% load i18n %}
 
  <div class="navigation-megamenu__col navigation-megamenu__col--teaser {{ class }}">
     <div class="m-highlight">
