+++ navigation/templates/navigation/blocks/nested_page_list.html
@@ -1,5 +1,3 @@
-{% load wagtail_modeltranslation %}
-
 <div class="{{ column_class }}">
     <ul class="navigation-megamenu__list">
         {% for group in col.groups %}
