+++ navigation/templates/navigation/blocks/page_list.html
@@ -1,5 +1,3 @@
-{% load wagtail_modeltranslation %}
-
 <div class="{{ column_class }}">
     {% if col.title or col.description %}
          <div class="m-highlight m-highlight--alt">
