+++ navigation/templates/navigation/blocks/type_a.html
@@ -1,4 +1,4 @@
-{% load wagtailcore_tags wagtail_modeltranslation %}
+{% load wagtailcore_tags %}
 
 <div class="navigation-megamenu">
     {% include_block value.highlight %}
